[squash] re: fix .Net timeout argument.

+++ gremlin-dotnet/test/Gremlin.Net.IntegrationTest/Driver/MessagesTests.cs
@@ -101,8 +101,8 @@ namespace Gremlin.Net.IntegrationTest.Driver
             var gremlinServer = new GremlinServer(TestHost, TestPort);
             using (var gremlinClient = new GremlinClient(gremlinServer))
             {
-                const int timeOutInMs = 1;
-                const long scriptSleepTimeInMs = 5000;
+                const long timeOutInMs = 1L;
+                const int scriptSleepTimeInMs = 5000;
                 var sleepScript = _requestMessageProvider.GetSleepGremlinScript(scriptSleepTimeInMs);
 
                 var requestMsg =
